fix(tools): map Timing Opt configs and split sizer tests
Register Timing Opt/sizer workspace configs so `ecc config` can inspect
the db and DreamPlace files the inner legalize path actually reads.
Split the oversized sizer runner tests into runner, cleanup, and
EngineFlow modules.
