+++ chipcompiler/data/workspace/__init__.py
@@ -182,6 +182,7 @@ _STEP_CONFIG_KEYS: Final[dict[tuple[StepEnum, str], tuple[str, ...]]] = {
     (StepEnum.STA, "ecc"): ("db", StepEnum.RCX.value, StepEnum.STA.value),
     (StepEnum.PLACEMENT, "dreamplace"): ("dreamplace",),
     (StepEnum.LEGALIZATION, "dreamplace"): ("dreamplace",),
+    (StepEnum.TIMING_OPT, "sizer"): ("db", "dreamplace"),
 }
